feat(PrintGameInfo): 에러 출력문 리팩토링 및 로또 갯수 출력 함수 추가

+++ src/main/java/lotto/LottoList.java
@@ -18,10 +18,10 @@ public class LottoList {
         this.lottoCount = lottoAmount.divide(BigInteger.valueOf(AMOUNT_UNIT));
     }
 
-    public Long getLottoAmount() {
+    public BigInteger getLottoAmount() {
         return lottoAmount;
     }
-    public Long getLottoCount() {
+    public BigInteger getLottoCount() {
         return lottoCount;
     }
